requirements.txt
# LangChain — núcleo e integrações (v0.3+ LCEL)
langchain>=0.3.20
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langchain-groq>=0.2.0
langchain-huggingface>=0.1.0
langchain-community>=0.3.20
langchain-text-splitters>=0.3.0
langchain-qdrant>=0.2.0

# LangGraph — orquestração do pipeline RAG
langgraph>=0.4.0

# Vector store — Qdrant in-memory (sem servidor extra em dev)
# Em deploy real: QdrantClient(url="http://localhost:6333")
qdrant-client>=1.12.0

# Embeddings locais (sem custo por query)
sentence-transformers>=3.0.0

# Hybrid retrieval — BM25 keyword search
rank-bm25>=0.2.2

# PDF loader (PyMuPDFLoader)
pymupdf>=1.24

# Re-ranking — cross-encoder (opcional; fallback gracioso se ausente)
# pip install flashrank
# flashrank>=0.2.9

# API REST com streaming
fastapi>=0.115.0
uvicorn>=0.30.0

# Demo Gradio (Hugging Face Spaces) — gradio 6+ usa messages-format por padrão
gradio>=6.0.0

# Observabilidade — LangSmith (integração nativa LangGraph)
langsmith>=0.1.0
# Langfuse (alternativa open-source self-hostable)
# langfuse>=2.0.0

# Utilitários
tiktoken>=0.9.0
python-dotenv>=1.1.0
openai>=1.60.0
anthropic>=0.40.0
